+++ frontend/src/App.jsx
@@ -7,18 +7,17 @@ import TutorProfile from "./pages/TutorProfile";
 import SignUp from "./pages/auth/SignUp";
 import LogIn from "./pages/auth/LogIn";
 
-
 function App() {
   return (
-      <Router>
-        <Header/>
-        <Routes>
-          <Route path="/" element={<LandingPage />} />
-          <Route path="/tutor/profile/8" element={<TutorProfile />} />
-          <Route path="/signup" element={<SignUp />} />
-          <Route path="/login" element={<LogIn />} />
-        </Routes>
-      </Router>
+    <Router>
+      <Header />
+      <Routes>
+        <Route path="/" element={<LandingPage />} />
+        <Route path="/tutor/profile/25" element={<TutorProfile />} />
+        <Route path="/signup" element={<SignUp />} />
+        <Route path="/login" element={<LogIn />} />
+      </Routes>
+    </Router>
   );
 }
