+++ frontend/src/pages/TutorProfile.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { useParams } from "react-router-dom";
 import axios from "axios";
 import { API_ROUTES } from "../constant/APIRoutes";
-import { toast, Toaster } from 'react-hot-toast'
+import { toast, Toaster } from "react-hot-toast";
 
 import BasicInfor from "../components/tutor_profile/watch/BasicInfor";
 import BasicInforEdit from "../components/tutor_profile/edit/BasicInforEdit";
@@ -16,9 +16,9 @@ import Awards from "../components/tutor_profile/watch/Award";
 import AwardsEdit from "../components/tutor_profile/edit/AwardEdit";
 
 const TutorProfile = () => {
-  const userId = 8; // fix later
   const [loggedInUserId, setLoggedInUserId] = useState(null); // Store logged-in user's ID
   const [profileData, setProfileData] = useState(null);
+  const [userId, setUserId] = useState(localStorage.getItem("userId"));
 
   useEffect(() => {
     // Get logged-in user's ID from localStorage or your auth provider
@@ -48,27 +48,32 @@ const TutorProfile = () => {
     );
   }
 
-
   // Determine if the logged-in user is the profile owner
   const isOwner = String(loggedInUserId) === String(userId);
 
-  const basicprofileData = profileData.profile
+  const basicprofileData = profileData.profile;
   const educationData = profileData.education;
   const coursesData = profileData.course;
   const examsData = profileData.exam;
   const awardsData = profileData.award;
 
   const showToast = (status) => {
-    if (status === 'success') {
-      toast.success('Updated Successfully', { duration: 1500 })
-    } else if (status === 'username_exists') {
-      toast.error('This username already exists. Please try a different one.', { duration: 1500 })
-    } else if (status === 'email_exists') {
-      toast.error('This email already exists. Please try a different one.', { duration: 1500 })
+    if (status === "success") {
+      toast.success("Updated Successfully", { duration: 1500 });
+    } else if (status === "username_exists") {
+      toast.error("This username already exists. Please try a different one.", {
+        duration: 1500,
+      });
+    } else if (status === "email_exists") {
+      toast.error("This email already exists. Please try a different one.", {
+        duration: 1500,
+      });
     } else {
-      toast.error('Failed to save changes. Please try again.', { duration: 1500 })
+      toast.error("Failed to save changes. Please try again.", {
+        duration: 1500,
+      });
     }
-  }
+  };
 
   return (
     <div className="bg-yellow-50 h-auto px-24 py-1 space-y-6 pb-8">
@@ -76,7 +81,11 @@ const TutorProfile = () => {
       {/* Basic Information */}
       <div className="rounded-xl h-auto bg-white">
         {isOwner ? (
-          <BasicInforEdit data={basicprofileData} userId={userId} showToast={showToast} />
+          <BasicInforEdit
+            data={basicprofileData}
+            userId={userId}
+            showToast={showToast}
+          />
         ) : (
           <BasicInfor data={basicprofileData} />
         )}
@@ -85,7 +94,11 @@ const TutorProfile = () => {
       {/* Education Section */}
       <div className="rounded-xl h-auto bg-white px-10 py-8">
         {isOwner ? (
-          <EducationEdit data={educationData} userId={userId} showToast={showToast} />
+          <EducationEdit
+            data={educationData}
+            userId={userId}
+            showToast={showToast}
+          />
         ) : (
           <Education data={educationData} />
         )}
@@ -94,7 +107,11 @@ const TutorProfile = () => {
       {/* Courses Section */}
       <div className="rounded-xl h-auto bg-white px-10 py-8">
         {isOwner ? (
-          <CoursesEdit data={coursesData} userId={userId} showToast={showToast} />
+          <CoursesEdit
+            data={coursesData}
+            userId={userId}
+            showToast={showToast}
+          />
         ) : (
           <Courses data={coursesData} />
         )}
