+++ frontend/src/services/authService.js
@@ -3,8 +3,6 @@ import axios from "axios";
 const API_URL = "http://127.0.0.1:8000/auth/";
 
 export const register = async (userData) => {
-  console.log(userData);
-
   const response = await axios.post(`${API_URL}register/`, userData);
   return response.data;
 };
@@ -12,11 +10,12 @@ export const register = async (userData) => {
 export const login = async (userData) => {
   const response = await axios.post(`${API_URL}login/`, userData);
   if (response.data.access) {
-    localStorage.setItem("user", JSON.stringify(response.data));
+    localStorage.setItem("userId", response.data.userId);
   }
   return response.data;
 };
 
 export const logout = () => {
-  localStorage.removeItem("user");
+  // localStorage.removeItem("user");
+  localStorage.clear();
 };
